TOS, Privacy, Bugfixes and Cleanup
