+++ package.json
@@ -18,12 +18,14 @@
     "vue": "^2.5.13",
     "vue-authenticate": "^1.3.4",
     "vue-axios": "^2.0.2",
-    "vue-notification": "^1.3.4",
+    "vue-cookies": "^1.5.4",
+    "vue-js-modal": "^1.3.9",
+    "vue-notification": "^1.3.6",
     "vue-router": "^3.0.1",
     "vuex": "^3.0.1"
   },
   "devDependencies": {
-    "autoprefixer": "^7.2.4",
+    "autoprefixer": "^7.2.5",
     "babel-core": "^6.22.1",
     "babel-loader": "^7.1.1",
     "babel-plugin-istanbul": "^4.1.1",
@@ -35,7 +37,7 @@
     "chalk": "^2.0.1",
     "copy-webpack-plugin": "^4.3.1",
     "cross-env": "^5.1.3",
-    "css-loader": "^0.28.0",
+    "css-loader": "^0.28.9",
     "eventsource-polyfill": "^0.9.6",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^1.1.6",
@@ -52,25 +54,25 @@
     "karma-spec-reporter": "0.0.31",
     "karma-webpack": "^2.0.9",
     "mocha": "^3.2.0",
-    "node-notifier": "^5.1.2",
+    "node-notifier": "^5.2.1",
     "optimize-css-assets-webpack-plugin": "^3.2.0",
     "ora": "^1.2.0",
     "phantomjs-prebuilt": "^2.1.14",
     "portfinder": "^1.0.13",
     "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.0.8",
+    "postcss-loader": "^2.0.10",
     "rimraf": "^2.6.0",
-    "semver": "^5.3.0",
+    "semver": "^5.5.0",
     "shelljs": "^0.7.6",
-    "sinon": "^4.1.3",
+    "sinon": "^4.2.2",
     "sinon-chai": "^2.8.0",
     "url-loader": "^0.5.8",
-    "vue-loader": "^13.6.2",
-    "vue-style-loader": "^3.0.1",
+    "vue-loader": "^13.7.1",
+    "vue-style-loader": "^3.1.2",
     "vue-template-compiler": "^2.5.13",
     "webpack": "^3.10.0",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "^2.9.7",
+    "webpack-bundle-analyzer": "^2.9.2",
+    "webpack-dev-server": "^2.11.1",
     "webpack-merge": "^4.1.0"
   },
   "engines": {
