+++ src/router/index.js
@@ -4,15 +4,16 @@ import Router from 'vue-router'
 import SignUp from '@/components/signup'
 import store from '../store/store'
 
-import pickAccount from '@/components/steps/pick_account'
-import saveKey from '@/components/steps/save_key'
-import verifyMail from '@/components/steps/email_verification'
-import verifyPhone from '@/components/steps/phone_verification'
-import confirmPhone from '@/components/steps/confirm_phone'
-import confirmEmail from '@/components/callbacks/confirm_email'
+import pickAccount from '@/components/steps/account/pick_account'
+import saveKey from '@/components/steps/account/save_key'
 
-import successEmail from '@/components/steps/success_email'
-import successPhone from '@/components/steps/success_phone'
+import verifyMail from '@/components/steps/email/email_verification'
+import successEmail from '@/components/steps/email/success_email'
+import confirmEmail from '@/components/steps/email/confirm_email'
+
+import confirmPhone from '@/components/steps/phone/confirm_phone'
+import verifyPhone from '@/components/steps/phone/phone_verification'
+import successPhone from '@/components/steps/phone/success_phone'
 
 import signInCallback from '@/components/callbacks/signin'
 import loginCallback from '@/components/callbacks/login'
