+++ src/store/store.js
@@ -4,6 +4,10 @@ import VueAxios from 'vue-axios'
 import VueAuthenticate from 'vue-authenticate'
 import axios from 'axios'
 import crypto from 'crypto'
+import VueCookies from 'vue-cookies'
+
+
+
 
 //import axios from 'axios'
 
@@ -13,6 +17,7 @@ const client = new dsteem.Client('https://api.steemit.com')
 
 Vue.use(Vuex)
 Vue.use(VueAxios, axios)
+Vue.use(VueCookies)
 
 var vueAuth = VueAuthenticate.factory(Vue.prototype.$http, {
     baseUrl: process.env.API_PATH,
@@ -23,6 +28,8 @@ var vueAuth = VueAuthenticate.factory(Vue.prototype.$http, {
       }
 })
 
+
+
 const store = new Vuex.Store({
     state: {
         verification_required: true,
@@ -30,12 +37,28 @@ const store = new Vuex.Store({
         chosen_account_name:'',
         chosen_email:'',
         generated_password: null,
-        current_user_object: {}
+        current_user_object: {},
+        chosen_phonenumber: '',
+        chosen_countrycode: ''
     },
     actions: {
+        checkCookie: ({ commit, state }, payload) => {
+            let tries = Number(window.$cookies.get(payload.cookie_name)) ? Number(window.$cookies.get(payload.cookie_name)) : 0
+            if(tries >= 3) { return false }
+            return true
+        },
+        changeCookie: ({ commit, state }, payload) => {
+            let tries = Number(window.$cookies.get(payload.cookie_name)) ? Number(window.$cookies.get(payload.cookie_name)) : 0
+            window.$cookies.set(payload.cookie_name, tries += 1, payload.timeout)
+        },
         createPassword: async ({ commit, state }, payload) => {
           let password = steem.formatter.createSuggestedPassword() // crypto.randomBytes(16).toString('hex')
           return commit('setGeneratedPassword', { password })
+        },
+        acceptModal: ({ commit, state}, payload) => {
+            return axios.post(`${process.env.API_PATH}/auth/account/accept`, { user_id: state.current_user_object._id, type: payload.type })
+                         .then(response => { return response })
+                         .catch(err => { console.log(err);return err.response ? err.response : err })
         },
         createAccount: async ({ commit, state }, payload) => {
           const owner_key = dsteem.PrivateKey.fromLogin(state.chosen_account_name, state.generated_password, 'owner')
@@ -75,7 +98,7 @@ const store = new Vuex.Store({
                         .catch(err => { return err.response ? err.response : err })
         },
         requestPhone: ({ commit, state }, payload) => {
-            return axios.post(`${process.env.API_PATH}/auth/phone/request`, { user_id: state.current_user_object._id, country_code: payload.country_code, phone_number: payload.phone_number })
+            return axios.post(`${process.env.API_PATH}/auth/phone/request`, { user_id: state.current_user_object._id, country_code: state.chosen_countrycode, phone_number: state.chosen_phonenumber })
                         .then(response => { return response })
                         .catch(err => { console.log(err); return err.response ? err.response : err })
         },
@@ -83,6 +106,16 @@ const store = new Vuex.Store({
             return axios.post(`${process.env.API_PATH}/auth/phone/confirm`, { user_id: state.current_user_object._id, code: payload.code })
                         .then(response => { return response })
                         .catch(err => { console.log(err);return err.response ? err.response : err })
+        },
+        resendPhone: ({ commit, state }, payload) => {
+            return  axios.post(`${process.env.API_PATH}/auth/phone/resend`, { user_id: state.current_user_object._id, country_code: state.chosen_countrycode, phone_number: state.chosen_phonenumber })
+                         .then(response => { return response })
+                         .catch(err => { console.log(err);return err.response ? err.response : err })
+        },
+        resetPendingPhone: ({ commit, state }, payload) => {
+            return axios.post(`${process.env.API_PATH}/auth/phone/reset`, { user_id: state.current_user_object._id })
+                         .then(response => { return response })
+                         .catch(err => { console.log(err);return err.response ? err.response : err })
         },
         login: ({ commit, state }, payload) => {
           let url = "https://v2.steemconnect.com/oauth2/authorize?client_id=utopian.app&response_type=code&redirect_uri=https%3A%2F%2Futopian.io%2Fcallback&scope=vote,comment,comment_delete,comment_options,custom_json,claim_reward_balance,offline"
@@ -93,7 +126,8 @@ const store = new Vuex.Store({
         setCurrentUserObject: (state, { object }) => { Vue.set(state, 'current_user_object', object) },
         setChosenAccName: (state, { name }) => { Vue.set(state, 'chosen_account_name', name) },
         setChosenEmail: (state, { email }) => { Vue.set(state, 'chosen_email', email) },
-        setGeneratedPassword: (state, { password }) => { Vue.set(state, 'generated_password', password) }
+        setGeneratedPassword: (state, { password }) => { Vue.set(state, 'generated_password', password) },
+        setPhoneNumber: (state, {country_code, phone_number}) => { Vue.set(state, 'chosen_countrycode', country_code); Vue.set(state, 'chosen_phonenumber', phone_number) }
     },
     getters: {
         currentUserObject: state => { return state.current_user_object },
